requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-appconf==1.0.4
django-datatables-view==1.19.1
django-environ==0.4.5
django-filter==2.3.0
django-gsheets==0.0.10
django-taggit==1.3.0
django-tailwind==0.9.0
djangorestframework==3.11.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.19.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
oauthlib==3.1.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
